conversational-image-chatbo/requirements.txt
transformers
torch
Pillow
opencv-python
gradio